app/requirements.txt
pymongo
requests
discord.py[voice]
daiquiri
